Volunteer page should function now, adding working App.js
